Task: added local scopes for imports to prevent circular dependency and added routes in routes.py
