+++ app/app.py
@@ -150,131 +150,5 @@ class AddressForm(FlaskForm):
     submit = SubmitField("Submit")
 
 
-
-@app.route('/', methods=['GET'])
-def index():
-    user = {'username': 'Address Project'}
-    all_addresses = Addresses.query.order_by(Addresses.id)
-    return render_template('index.html', title='Home', user=user, all_addresses = all_addresses)
-
-
-@app.route('/view/<int:address_id>', methods=['GET'])
-def record_view(address_id):
-    # print(db.session.query().filter_by(id=address_id).first())
-    print(Addresses.query.get(address_id).fname)
-    return render_template('view.html', title='View Form', city=Addresses.query.get(address_id))
-
-
-@app.route('/edit/<int:address_id>', methods=['GET'])
-def form_edit_get(address_id):
-    obj = Addresses.query.filter_by(id=address_id).one()
-    return render_template('edit.html', title='Edit Form', address=obj)
-#
-#
-
-
-@app.route('/edit/<int:address_id>', methods=['POST'])
-def form_update_post(address_id):
-    obj = Addresses.query.filter_by(id=address_id).one()
-    obj.fname = request.form.get('fname')
-    obj.lname = request.form.get('lname')
-    obj.address = request.form.get('address')
-    obj.city = request.form.get('city')
-    obj.state = request.form.get('state')
-    obj.zip_code = request.form.get('zip_code')
-    db.session.flush()
-    db.session.commit()
-    return redirect("/", code=302)
-#
-#
-
-
-@app.route('/address/new', methods=['POST'])
-def form_insert_get():
-        form = AddressForm()
-        addressNew = Addresses(fname=form.fname.data, lname=form.lname.data, address=form.address.data,
-                               city=form.city.data, state=form.state.data, zip_code=form.zip_code.data)
-        fname = form.fname
-        db.session.add(addressNew)
-        db.session.commit()
-        form.fname.data = ''
-        form.lname.data = ''
-        form.address.data = ''
-        form.city.data = ''
-        form.state.data = ''
-        form.zip_code.data = ''
-        all_addresses = Addresses.query.order_by(Addresses.id)
-        return render_template('new.html', title='New Address Form', form=form, fname=fname, all_addresses = all_addresses)
-
-
-@app.route('/address/new', methods=['GET'])
-def form_insert_post():
-    form = AddressForm()
-    all_addresses = Addresses.query.order_by(Addresses.id)
-    return render_template('new.html', title='New Address Form', form=form, all_addresses=all_addresses)
-    return redirect("/", code=302)
-
-
-@app.route('/delete/<int:address_id>', methods=['POST'])
-def form_delete_post(address_id):
-    obj = Addresses.query.filter_by(id=address_id).one()
-    db.session.delete(obj)
-    db.session.commit();
-    return redirect("/", code=302)
-
-
-@app.route('/api/v1/addresses', methods=['GET'])
-def api_browse() -> str:
-    resp = Addresses.query.all()
-    json_arr = []
-    for temp in resp:
-        json_arr.append(temp.toDict())
-    return jsonify(json_arr)
-
-
-@app.route('/api/v1/addresses/<int:address_id>', methods=['GET'])
-def api_retrieve(address_id) -> str:
-    resp = Addresses.query.filter_by(id=address_id).one()
-    return jsonify(resp.toDict())
-
-
-@app.route('/api/v1/addresses/<int:address_id>', methods=['PUT'])
-def api_edit(address_id) -> str:
-    content = request.json
-    obj = Addresses.query.filter_by(id=address_id).one()
-    obj.fname = content['fname']
-    obj.lname = content['lname']
-    obj.address = content['address']
-    obj.city = content['city']
-    obj.state = content['state']
-    obj.zip_code = content['zip_code']
-    db.session.commit()
-    resp = Response(status=200, mimetype='application/json')
-    return resp
-
-
-@app.route('/api/v1/addresses', methods=['POST'])
-def api_add() -> str:
-
-    content = request.json
-
-    newAddress = Addresses(fname = content['fname'], lname = content['lname'], address = content['address'],
-                 city = content['city'], state = content['state'],
-                 zip_code = content['zip_code'])
-    db.session.add(newAddress)
-    db.session.commit()
-    resp = Response(status=201, mimetype='application/json')
-    return resp
-
-
-@app.route('/api/v1/addresses/<int:address_id>', methods=['DELETE'])
-def api_delete(address_id) -> str:
-    obj = Addresses.query.filter_by(id=address_id).one()
-    db.session.delete(obj)
-    db.session.commit();
-    resp = Response(status=200, mimetype='application/json')
-    return resp
-
-
 if __name__ == '__main__':
     app.run(host="0.0.0.0")
